souphttpsrc: share the cookiejar in the context, not the cookies
The cookies can mutate at every request and it would need proper tracking
to keep updating the context and then copying it back to the soup cookie
jar. It is easier to let the cookiejar be shared and handle it directly there.

Other http source implementations can do the same with their cookiejar
implementations